refactor(client): Extract next steps from complete_sign_up into the brokers.
complete_sign_up was turning into a mess, and the CAD code was more difficult than
need be because of it. We had to enable CAD for sign up, but not sign in,
we had to redirect based on what type of verification the user was doing.

Brokers allow for many of these decisions to be made declaratively.

Add two new broker methods:
* afterCompleteAddSecondaryEmail
* afterCompleteSignIn